Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main(){
-    double num;
+    float num;
     scanf("%f",&num);
-    printf("%f",num);
+    printf("%.2f",num);
     return 0;
 }
